+++ src/navigation/index.ts
@@ -3,4 +3,5 @@ export * from "./router";
 export * from "./daily-target";
 export * from "./question-target";
 export * from "./web-target";
+export * from "./deep-link";
 export * from "./checkpoint-service";